backend/requirements.txt
<<<<<<< Updated upstream:backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
openai==1.3.5
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
=======
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.20.0,<0.25.0
gunicorn>=20.0.0,<22.0.0
openai>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=0.19.0,<2.0.0
python-multipart>=0.0.5,<1.0.0
>>>>>>> Stashed changes:requirements.txt
